Desktop/juris-nexus-poc/requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
ansi2html==1.9.2
anthropic==0.49.0
anyio==3.7.1
asttokens==3.0.0
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.2
blinker==1.9.0
blis==0.7.11
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
contourpy==1.3.1
coverage==7.7.1
cryptography==44.0.0
cycler==0.12.1
cymem==2.0.11
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.4.1
decorator==5.2.1
Deprecated==1.2.15
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fake-useragent==2.1.0
fastapi==0.115.11
filelock==3.13.1
Flask==2.2.5
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
future==1.0.0
gensim==4.3.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-aiplatform==1.36.4
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.71.0
grpcio-status==1.62.3
gunicorn==23.0.0
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.7
httpx==0.25.1
huggingface-hub==0.29.3
idna==3.10
iniconfig==2.1.0
ipython==8.34.0
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpickle==4.0.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
line-bot-sdk==3.16.2
loguru==0.7.3
lxml==5.3.1
Mako==1.3.9
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.7.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==2.2.4
openai==1.66.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==1.5.3
parso==0.8.4
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==11.1.0
plotly==5.15.0
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic_core==2.16.2
Pygments==2.19.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2024.2
pyvis==0.3.2
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rsa==4.9
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.0
seaborn==0.12.2
selenium==4.29.0
sentence-transformers==3.4.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
snownlp==0.12.3
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-pkuseg==0.0.33
SQLAlchemy==2.0.23
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
text2vec==1.3.3
textblob==0.17.1
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.9.4
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.24.0
vertexai==0.0.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
webdriver-manager==4.0.1
websocket-client==1.8.0
Werkzeug==2.2.3
wordcloud==1.9.2
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.2.1
xxhash==3.5.0
yarl==1.18.3
zh-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.7.0/zh_core_web_sm-3.7.0-py3-none-any.whl#sha256=f51075665749e07406d629d1055ce5a68635fae6ab3c34257ee798c62b4fc431
